refactor simulation data using streamlit-book
